Put entire timestamp in :before section
